+++ src/index.ts
@@ -1,2 +1,5 @@
 export type { Renderable } from './Renderable';
 export * as Renderer2D from './Renderer2D';
+export * from './Primitives';
+export * from './Color';
+export * from './Types';